src/Frontend/SFC.Frontend/src/app/core/components/header/services/header.service.ts
import { Inject, Injectable } from '@angular/core';
import { DOCUMENT, UIConstants } from 'ngx-sfc-common';
import { BehaviorSubject, Observable, tap } from 'rxjs';

@Injectable({
    providedIn: 'root'
})
export class HeaderService {
    public get open(): boolean { return this.openSubject.value; }

    private openSubject: BehaviorSubject<boolean> = new BehaviorSubject<boolean>(false);

    public open$: Observable<boolean> = this.openSubject.asObservable()
        .pipe(
            tap(value => this.document.body.style.overflow = value
                ? UIConstants.CSS_VISIBILITY_HIDDEN
                : UIConstants.CSS_INITIAL)
        );

    constructor(@Inject(DOCUMENT) private document: Document) { }

    public toggle(): void {
        this.openSubject.next(!this.open);
    }

    public toggleByValue(value: boolean): void {
        this.openSubject.next(value);
    }
}
